Add hooks for called CUDA API functions, renamed class.
git-svn-id: http://cuda-waste.googlecode.com/svn/trunk@16 b05b5d53-d61a-d14b-5d44-457121efbe5e

Former-commit-id: 55776f2448435fe9378c789975f53774c026db03

+++ ptxp/driver.cpp
@@ -9,7 +9,7 @@ static void print(pANTLR3_BASE_TREE node, int level)
 	for (int i = 0; i < level; ++i)
 	    printf("   ");
 	printf("%s\n", node->getText(node)->chars);
-	for (int i = 0; i < node->getChildCount(node); ++i)
+	for (int i = 0; i < (int)node->getChildCount(node); ++i)
 	{
 	    pANTLR3_BASE_TREE child = (pANTLR3_BASE_TREE)node->getChild(node, i);
 	    print(child, level+1);
